fix no lint in files
