+++ cmd/system-probe/modules/network_tracer.go
@@ -9,7 +9,6 @@ package modules
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"net/http"
@@ -272,19 +271,6 @@ func writeConnections(w http.ResponseWriter, marshaler marshal.Marshaler, cs *ne
 	}
 
 	log.Tracef("/connections: %d connections", len(cs.Conns))
-}
-
-// writeDisabledProtocolMessage is used by USM endpoint handlers on Linux.
-//
-//nolint:unused
-func writeDisabledProtocolMessage(protocolName string, w http.ResponseWriter) {
-	log.Warnf("%s monitoring is disabled", protocolName)
-	w.WriteHeader(404)
-	// Writing JSON to ensure compatibility when using the jq bash utility for output
-	outputString := map[string]string{"error": protocolName + " monitoring is disabled"}
-	// We are marshaling a static string, so we can ignore the error
-	buf, _ := json.Marshal(outputString)
-	w.Write(buf)
 }
 
 func writeConntrackTable(table *tracer.DebugConntrackTable, w http.ResponseWriter) {
