+++ pkg/network/tracer/connection/ebpf_tracer_unsupported.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2024-present Datadog, Inc.
 
-//go:build !linux || !linux_bpf
+//go:build !linux_bpf && !darwin
 
 // Package connection provides network connection tracking functionality.
 package connection
@@ -16,8 +16,6 @@ import (
 )
 
 // newEbpfTracer is not supported on non-Linux platforms
-//
-//nolint:unused // used as stub for connection.NewTracer on non-Linux
 func newEbpfTracer(_ *config.Config, _ telemetry.Component) (Tracer, error) {
 	return nil, errors.New("eBPF tracer not supported on this platform")
 }
