+++ pkg/network/tracer/connection/tracer.go
@@ -32,12 +32,6 @@ const (
 	TracerTypeDarwin
 )
 
-const (
-	// maxActive configures the maximum number of instances of the kretprobe-probed functions handled simultaneously.
-	// This value should be enough for typical workloads (e.g. some amount of processes blocked on the `accept` syscall).
-	maxActive = 512 //nolint:unused // used by Linux eBPF tracer
-)
-
 // Tracer is the common interface implemented by all connection tracers.
 type Tracer interface {
 	// Start begins collecting network connection data.
